Move internal Field components to co-located internal-only file
